// src/app/shared/models/statistics.model.ts
export interface Statistics {
  year?: number;
  month?: number;
  day?: number;
  caloriesGoal?: number;
  proteinGoal?: number;
  fatsGoal?: number;
  carbohydratesGoal?: number;
  caloriesResult?: number;
  proteinResult?: number;
  fatsResult?: number;
  carbohydratesResult?: number;
  totalScore?: number;
}